+++ R/widget.R
@@ -1,22 +1,24 @@
 sas_widget <- function(
   lst,
   log,
-  width = NULL,
-  height = NULL,
+  capture,
+  height = "auto",
+  width = "auto",
   elementId = NULL
 ) {
   # forward options using x
   x <- list(
     lst = lst,
-    log = log
+    log = log,
+    capture = capture
   )
 
   # create widget
   htmlwidgets::createWidget(
     name = 'sas_widget',
     x,
-    width = "auto",
-    height = "auto",
+    height = height,
+    width = width,
     sizingPolicy = htmlwidgets::sizingPolicy(
       padding = 0,
       viewer.fill = TRUE,